+++ crates/stackable-webhook/src/lib.rs
@@ -1,6 +1,6 @@
 //! Utility types and functions to easily create ready-to-use webhook servers
 //! which can handle different tasks, for example CRD conversions. All webhook
-//! servers use HTTPS by defaultThis library is fully compatible with the
+//! servers use HTTPS by default. This library is fully compatible with the
 //! [`tracing`] crate and emits debug level tracing data.
 //!
 //! Most users will only use the top-level exported generic [`WebhookServer`]
@@ -8,27 +8,34 @@
 //! routes and their handler functions.
 //!
 //! ```
-//! use stackable_webhook::{WebhookServer, Options};
+//! use stackable_webhook::{WebhookServer, WebhookOptions};
 //! use axum::Router;
 //!
+//! # async fn test() {
 //! let router = Router::new();
-//! let server = WebhookServer::new(router, Options::default());
+//! let (server, cert_rx) = WebhookServer::new(router, WebhookOptions::default())
+//!     .await
+//!     .expect("failed to create WebhookServer");
+//! # }
 //! ```
 //!
 //! For some usages, complete end-to-end [`WebhookServer`] implementations
-//! exist. One such implementation is the [`ConversionWebhookServer`][1]. The
-//! only required parameters are a conversion handler function and [`Options`].
+//! exist. One such implementation is the [`ConversionWebhookServer`][1].
 //!
 //! This library additionally also exposes lower-level structs and functions to
-//! enable complete controll over these details if needed.
+//! enable complete control over these details if needed.
 //!
 //! [1]: crate::servers::ConversionWebhookServer
 use axum::{Router, routing::get};
 use futures_util::{FutureExt as _, pin_mut, select};
 use snafu::{ResultExt, Snafu};
 use stackable_telemetry::AxumTraceLayer;
-use tokio::signal::unix::{SignalKind, signal};
+use tokio::{
+    signal::unix::{SignalKind, signal},
+    sync::mpsc,
+};
 use tower::ServiceBuilder;
+use x509_cert::Certificate;
 
 // use tower_http::trace::TraceLayer;
 use crate::tls::TlsServer;
@@ -39,11 +46,7 @@ pub mod servers;
 pub mod tls;
 
 // Selected re-exports
-pub use crate::options::Options;
-
-/// A result type alias with the library-level [`Error`] type as teh default
-/// error type.
-pub type Result<T, E = Error> = std::result::Result<T, E>;
+pub use crate::options::WebhookOptions;
 
 /// A generic webhook handler receiving a request and sending back a response.
 ///
@@ -56,25 +59,16 @@ pub trait WebhookHandler<Req, Res> {
     fn call(self, req: Req) -> Res;
 }
 
-/// A generic webhook handler receiving a request and state and sending back
-/// a response.
-///
-/// This trait is not intended to be  implemented by external crates and this
-/// library provides various ready-to-use implementations for it. One such an
-/// implementation is part of the [`ConversionWebhookServer`][1].
-///
-/// [1]: crate::servers::ConversionWebhookServer
-pub trait StatefulWebhookHandler<Req, Res, S> {
-    fn call(self, req: Req, state: S) -> Res;
-}
+/// A result type alias with the [`WebhookError`] type as the default error type.
+pub type Result<T, E = WebhookError> = std::result::Result<T, E>;
 
 #[derive(Debug, Snafu)]
-pub enum Error {
+pub enum WebhookError {
     #[snafu(display("failed to create TLS server"))]
-    CreateTlsServer { source: tls::Error },
+    CreateTlsServer { source: tls::TlsServerError },
 
     #[snafu(display("failed to run TLS server"))]
-    RunTlsServer { source: tls::Error },
+    RunTlsServer { source: tls::TlsServerError },
 }
 
 /// A ready-to-use webhook server.
@@ -88,17 +82,16 @@ pub enum Error {
 ///
 /// [1]: crate::servers::ConversionWebhookServer
 pub struct WebhookServer {
-    options: Options,
-    router: Router,
+    tls_server: TlsServer,
 }
 
 impl WebhookServer {
     /// Creates a new ready-to-use webhook server.
     ///
-    /// The server listens on `socket_addr` which is provided via the [`Options`]
-    /// and handles routing based on the provided Axum `router`. Most of the time
-    /// it is sufficient to use [`Options::default()`]. See the documentation
-    /// for [`Options`] for more details on the default values.
+    /// The server listens on `socket_addr` which is provided via the [`WebhookOptions`] and handles
+    /// routing based on the provided Axum `router`. Most of the time it is sufficient to use
+    /// [`WebhookOptions::default()`]. See the documentation for [`WebhookOptions`] for more details
+    /// on the default values.
     ///
     /// To start the server, use the [`WebhookServer::run()`] function. This will
     /// run the server using the Tokio runtime until it is terminated.
@@ -106,29 +99,66 @@ impl WebhookServer {
     /// ### Basic Example
     ///
     /// ```
-    /// use stackable_webhook::{WebhookServer, Options};
+    /// use stackable_webhook::{WebhookServer, WebhookOptions};
     /// use axum::Router;
     ///
+    /// # async fn test() {
     /// let router = Router::new();
-    /// let server = WebhookServer::new(router, Options::default());
+    /// let (server, cert_rx) = WebhookServer::new(router, WebhookOptions::default())
+    ///     .await
+    ///     .expect("failed to create WebhookServer");
+    /// # }
     /// ```
     ///
     /// ### Example with Custom Options
     ///
     /// ```
-    /// use stackable_webhook::{WebhookServer, Options};
+    /// use stackable_webhook::{WebhookServer, WebhookOptions};
     /// use axum::Router;
     ///
-    /// let options = Options::builder()
+    /// # async fn test() {
+    /// let options = WebhookOptions::builder()
     ///     .bind_address([127, 0, 0, 1], 8080)
+    ///     .add_subject_alterative_dns_name("my-san-entry")
     ///     .build();
     ///
     /// let router = Router::new();
-    /// let server = WebhookServer::new(router, options);
+    /// let (server, cert_rx) = WebhookServer::new(router, options)
+    ///     .await
+    ///     .expect("failed to create WebhookServer");
+    /// # }
     /// ```
-    pub fn new(router: Router, options: Options) -> Self {
+    pub async fn new(
+        router: Router,
+        options: WebhookOptions,
+    ) -> Result<(Self, mpsc::Receiver<Certificate>)> {
         tracing::trace!("create new webhook server");
-        Self { options, router }
+
+        // TODO (@Techassi): Make opt-in configurable from the outside
+        // Create an OpenTelemetry tracing layer
+        tracing::trace!("create tracing service (layer)");
+        let trace_layer = AxumTraceLayer::new().with_opt_in();
+
+        // Use a service builder to provide multiple layers at once. Recommended
+        // by the Axum project.
+        //
+        // See https://docs.rs/axum/latest/axum/middleware/index.html#applying-multiple-middleware
+        // TODO (@NickLarsenNZ): rename this server_builder and keep it specific to tracing, since it's placement in the chain is important
+        let service_builder = ServiceBuilder::new().layer(trace_layer);
+
+        // Create the root router and merge the provided router into it.
+        tracing::debug!("create core router and merge provided router");
+        let router = router
+            .layer(service_builder)
+            // The health route is below the AxumTraceLayer so as not to be instrumented
+            .route("/health", get(|| async { "ok" }));
+
+        tracing::debug!("create TLS server");
+        let (tls_server, cert_rx) = TlsServer::new(router, options)
+            .await
+            .context(CreateTlsServerSnafu)?;
+
+        Ok((Self { tls_server }, cert_rx))
     }
 
     /// Runs the Webhook server and sets up signal handlers for shutting down.
@@ -170,33 +200,6 @@ impl WebhookServer {
     async fn run_server(self) -> Result<()> {
         tracing::debug!("run webhook server");
 
-        // TODO (@Techassi): Make opt-in configurable from the outside
-        // Create an OpenTelemetry tracing layer
-        tracing::trace!("create tracing service (layer)");
-        let trace_layer = AxumTraceLayer::new().with_opt_in();
-
-        // Use a service builder to provide multiple layers at once. Recommended
-        // by the Axum project.
-        //
-        // See https://docs.rs/axum/latest/axum/middleware/index.html#applying-multiple-middleware
-        // TODO (@NickLarsenNZ): rename this server_builder and keep it specific to tracing, since it's placement in the chain is important
-        let service_builder = ServiceBuilder::new().layer(trace_layer);
-
-        // Create the root router and merge the provided router into it.
-        tracing::debug!("create core router and merge provided router");
-        let router = self
-            .router
-            .layer(service_builder)
-            // The health route is below the AxumTraceLayer so as not to be instrumented
-            .route("/health", get(|| async { "ok" }));
-
-        // Create server for TLS termination
-        tracing::debug!("create TLS server");
-        let tls_server = TlsServer::new(self.options.socket_addr, router)
-            .await
-            .context(CreateTlsServerSnafu)?;
-
-        tracing::info!("running TLS server");
-        tls_server.run().await.context(RunTlsServerSnafu)
+        self.tls_server.run().await.context(RunTlsServerSnafu)
     }
 }